backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-dotenv
pdfplumber
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-huggingface
langchain-groq
langchain-google-genai
sentence-transformers
faiss-cpu
pydantic>=2.0
aiofiles
httpx
